update for snapshot
